pyproject.toml
[project]
name = "alip"
version = "0.2.1"
description = "AI-Assisted Legacy Intelligence Platform - Trust-first legacy system analysis"
authors = [{name = "ALIP Team"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0.0",
    "anthropic>=0.18.0",
    "click>=8.1.0",
    "pyyaml>=6.0",
    "networkx>=3.0",
    "gitpython>=3.1.0",
    "sqlparse>=0.4.0",
    "PyPDF2>=3.0.0",
    "python-docx>=1.0.0",
    "jinja2>=3.1.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
alip = "alip.cli:main"

[build-system]
requires = ["setuptools>=65.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --cov=core --cov=skills --cov=agents --cov-report=term-missing"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
